sslman creation and terminal improvements/fixes
created the sslman plugin for creating easy ssl connections. This plugin is very WIP.

The terminal plugin had some minor updates. The code was improved to better follow style guidelines (function names have remained unchanged for now.) A stop command was added to facilitate clean exits, not just emergency stops. Also a bug was fixed which would cause the program to crash if there was whitespace at the end of the command entry.